LayoutTests/fast/shadow-dom/Element-slot-unscopable.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that Element.slot is [Unscopable].");

with (document.body) {
    try {
        slot;
        testFailed("Element.body is not [Unscopable]");
    } catch(e) {
        exception = e;
        shouldBeEqualToString("exception.name", "ReferenceError");
    }
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
