LayoutTests/fast/shadow-dom/Element-slot-unscopable-expected.txt

Tests that Element.slot is [Unscopable].

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS exception.name is "ReferenceError"
PASS successfullyParsed is true

TEST COMPLETE
